configure via cli/gui, cleanup, fixes, docs

+++ release.ps1
@@ -1,4 +1,4 @@
-# usage: release.ps1 <version>
+# usage: release.ps1 -Version=<version>
 Param
     (
         [parameter(Position=0,Mandatory=$true)]
@@ -10,32 +10,22 @@ $operatingsystem = @{
     Linux = "linux-x64"
     MacOS = "osx-x64"
 }
-$guiframework = @{
-    Windows = "Wpf"
-    Linux = "Gtk"
-    MacOS = "Mac"
-}
-$guinet = @{
-    Windows = "net8.0-windows"
-    Linux = "net8.0"
-    MacOS = "net8.0"
-}
 # create output folder if it doesn't exist
 $outputFolder = "release"
-Remove-Item -Force -Recurse -Path .\$outputFolder\
+if(Test-Path -Path $outputFolder)
+{
+    # remove folder if it exists already for new release
+    Remove-Item -Force -Recurse -Path .\$outputFolder\
+}
 New-Item -ItemType Directory -Force -Path .\$outputFolder\
 
-# CLI
 $operatingsystem.GetEnumerator() | ForEach-Object{
     $rid = $_.Value;
     $name = $_.Key;
+    # CLI
     dotnet publish .\CLI\CLI.csproj --runtime $rid --configuration Release --self-contained=true
     Compress-Archive -Path .\CLI\bin\Release\net8.0\$rid\publish\* -DestinationPath .\$outputFolder\VelvetBeautifier.CLI.$Version.$name.zip
-}
-# GUI
-$operatingsystem.GetEnumerator() | ForEach-Object{
-    $rid = $_.Value;
-    $name = $_.Key;
+    # GUI
     dotnet publish .\GUI\GUI.csproj --runtime $rid --configuration Release --self-contained=true
     Compress-Archive -Path .\GUI\bin\Release\net8.0\$rid\publish\* -DestinationPath .\$outputFolder\VelvetBeautifier.GUI.$Version.$name.zip
 }
